+++ packages/legacy/tools/src/models/index.ts
@@ -1,0 +1,16 @@
+import Package from './package';
+import PackageFile from './package-file';
+
+export type {
+  PackageBuildConfig,
+  PackageData,
+} from './package';
+
+export type {
+  PackageFileConfig,
+} from './package-file';
+
+export {
+  Package,
+  PackageFile,
+};